bebob: remove module parameters to reduce maintenance work
ALSA bebob driver has some parameters; 'index', 'id', 'enable'. Each
parameter consists of array. When a bebob unit is probed, it consumes
a set of elements from head of the array. Users can set preferred value to
the array, to assign these values to a card relevant to the unit.

The meaning of each parameter is:
 - index: used as 'card number' maintained by 'snd' module. The scope of
   card number is system wide. When duplicated value is set, the module
   detect it and the card fails to be probed.
 - id:  used as 'identification name'. For example, users can see it in
   '/proc/asound/cards'
 - enable: whether the element of array can be used or unused for a sound
   card instance. If not enabled, it skips and next element is used.

Well, this commit is going to purge these parameters. Most users have one
unit, therefore they don't require the parameters. If using several units,
they can know which card corresponds to a unit, because units on IEEE 1394
has global unique ID (GUID) and drivers in ALSA firewire stack except for
dice driver pushes the GUID to 'longname' field of sound card. As another
way, hwdep interface give a way for users to get the correspondence between
the sound card and GUID. For example, libhinawa is such an application.

Removal of the parameters reduces relevant codes, thus reduces maintenance
work. This commit purge them.