firewire-lib: cancel processing packets in struct snd_pcm_ops.pointer() call
In current implementation of IEC 61883-1/6 engine in ALSA firewire
stack, there's two chances to handle isochronous packets in process
contexts. One is a callback of struct snd_pcm_ops.pointer() and another
is the one of .ack(). When alsa-lib applications queue/dequeue PCM frames
according to mmap scenario, the isochronous packets are handled in below
call graph.

while ()
  ->poll()
  ->snd_pcm_hwsync()
    ->ioctl(HWSYNC)
      ->struct snd_pcm_ops.pointer()
      = amdtp_stream_pcm_pointer()
          ->fw_iso_context_flush_completions()
  ->snd_pcm_mmap_begin()
  (queue/dequeue PCM frames here.)
  ->snd_pcm_mmap_commit()
    ->ioctl(SYNC_PTR)
      ->struct snd_pcm_ops.ack()
      = amdtp_stream_pcm_ack()
        ->fw_iso_context_flush_completions()

In an above iteration, two calls of fw_iso_context_flush_completions()
are executed with quite small interval, thus few packets can be handled
in a call of .ack(). This is inefficient.

Although .ack() ensures to be called in process contexts, .pointer() is
called in interrupt contexts as well as process context. Especially,
this engine is designed to call snd_pcm_period_elapsed() in software IRQ
context of tasklet, scheduled in both of process context and interrupt
context of 1394 OHCI IR/IT context. To reduce closest two call of
.pointer(), a condition to judge running context is used in .pointer.
An idea to flush packets in a callback of .pointer() requires the above
extra care.

On the other hand, a callback of .ack() has no care like that, except
for being called within acquired lock. Therefore it's reasonable to
cancel .pointer() side.

This commit cancels flushing packets in a callback of .pointer(). The
callback of .pointer just returns the number of PCM frames in
intermediate PCM buffer or SNDRV_PCM_POS_XRUN. As a result:

while ()
  ->poll()
  ->snd_pcm_hwsync()
    ->ioctl(HWSYNC)
      ->struct snd_pcm_ops.pointer()
      = amdtp_stream_pcm_pointer()
  ->snd_pcm_mmap_begin()
  (queue/dequeue PCM frames here.)
  ->snd_pcm_mmap_commit()
    ->ioctl(SYNC_PTR)
      ->struct snd_pcm_ops.ack()
      = amdtp_stream_pcm_ack()
        ->fw_iso_context_flush_completions()